feat: implements anonymous user limit and login
Adds a limit of 3 roasts for anonymous users before
prompting them to log in. Implements a login modal to
handle this. Migrates anonymous roasts to user accounts
after login.

This change restricts the usage of the roast generation
feature to a limited number of attempts for unauthenticated
users, encouraging them to create an account for
unlimited usage.